feat(AAS-14): cocojson download
