+++ frontend/src/pages/AnnotateWorkspace.tsx
@@ -4,8 +4,6 @@ import { ToolSelectorSidebar } from '@/components/workspace/ToolSelectorSidebar'
 import { useState } from 'react'
 import { useGeneratePresignedDownloadUrlForImage } from '../../generated'
 import { Canvas } from '@/components/canvas'
-import { Button } from '@/components/ui/button'
-import CanvasExport from '@/components/canvas/CanvasExport'
 
 export function AnnotateWorkspace() {
   const { workspaceId, imageId } = useParams({ from: Route.id })
@@ -16,35 +14,20 @@ export function AnnotateWorkspace() {
   })
   const image = data?.data
 
-  const handleDownload = () => {
-    if (!canvasStateRef.current) {
-      alert('No canvas state found.')
-      return
-    }
-    const state = canvasStateRef.current.getState()
-    const json = CanvasExport(state, 'COCOJSON')
-    const blob = new Blob([json], { type: 'application/json' })
-    const url = URL.createObjectURL(blob)
-    const link = document.createElement('a')
-    link.href = url
-    link.download = `${image?.fileName.replace(/\.[^/.]+$/, '') || 'annotations'}.json`
-    document.body.appendChild(link)
-    link.click()
-    document.body.removeChild(link)
-
-    // Cleanup
-    URL.revokeObjectURL(url)
-  }
-
   return (
     <div className="flex h-screen items-center justify-center">
       <div className="relative flex flex-col items-center">
         <div className="flex flex-col items-center gap-4">
           <p>Currently editing: {image?.fileName}</p>
           {image?.url && (
-            <Canvas image={image.url} tool={active} workspaceId={workspaceId} imageId={imageId} />
+            <Canvas
+              image={image.url}
+              tool={active}
+              workspaceId={workspaceId}
+              imageId={imageId}
+              imageName={image.fileName}
+            />
           )}
-          <Button onClick={handleDownload}>Save Annotations</Button>
         </div>
         <div className="absolute left-full top-10 ml-2 flex flex-col justify-center">
           <ToolSelectorSidebar active={active} onSelect={setActive} />
